requirements.txt.txt
python-telegram-bot==20.7
APScheduler==3.10.4
Flask==2.3.3